django_app/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-note-generator-django"
version = "0.1.0"
description = "Django API for the medical note generator."
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<6.0",
  "djangorestframework>=3.15,<4.0",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
  "requests>=2.33.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-django>=4.9,<5.0",
]

[tool.setuptools.packages.find]
include = ["config*", "apps*"]
